requirements.txt
Pillow==11.2.1
streamlit==1.45.0
